build: rename android package
Renames android package from default `com.reactlibrary` to `org.hyperledger.indy.sdk.reactnative`
