+++ android/build.gradle
@@ -100,7 +100,7 @@ def configureReactNativePom(def pom) {
         name packageJson.title
         artifactId packageJson.name
         version = packageJson.version
-        group = "com.reactlibrary"
+        group = "org.hyperledger.indy.sdk.reactnative"
         description packageJson.description
         url packageJson.repository.baseUrl
